Extract remaining cloud sync domains
