+++ src/utils/sync/engine/domains/filesMetadataSyncDomain.ts
@@ -1,0 +1,76 @@
+import { STORES } from "@/utils/indexedDB";
+import { useCloudSyncStore } from "@/stores/useCloudSyncStore";
+import { useFilesStore, type FileSystemItem } from "@/stores/useFilesStore";
+import {
+  mergeDeletionMarkerMaps,
+  normalizeDeletionMarkerMap,
+  type DeletionMarkerMap,
+} from "@/utils/cloudSyncDeletionMarkers";
+import {
+  mergeFilesMetadataSnapshots,
+  type FilesMetadataSyncSnapshot,
+} from "@/utils/cloudSyncFileMerge";
+import {
+  applyIndexedDbStoreSnapshot,
+  serializeIndexedDbStoreSnapshot,
+  type FilesStoreSnapshotData,
+} from "@/utils/sync/engine/domains/indexedDbStoreSync";
+
+export interface FilesMetadataSnapshotData {
+  items: Record<string, FileSystemItem>;
+  libraryState: "uninitialized" | "loaded" | "cleared";
+  documents?: FilesStoreSnapshotData;
+  deletedPaths?: DeletionMarkerMap;
+}
+
+export async function serializeFilesMetadataSyncSnapshot(): Promise<FilesMetadataSnapshotData> {
+  const filesState = useFilesStore.getState();
+  const deletionMarkers = useCloudSyncStore.getState().deletionMarkers;
+
+  return {
+    items: filesState.items,
+    libraryState: filesState.libraryState,
+    documents: await serializeIndexedDbStoreSnapshot(STORES.DOCUMENTS),
+    deletedPaths: deletionMarkers.fileMetadataPaths,
+  };
+}
+
+export async function applyFilesMetadataSyncSnapshot(
+  data: FilesMetadataSnapshotData
+): Promise<void> {
+  const remoteDeletedPaths = normalizeDeletionMarkerMap(data.deletedPaths);
+  const cloudSyncState = useCloudSyncStore.getState();
+  const localDeletedPaths = cloudSyncState.deletionMarkers.fileMetadataPaths;
+  const localSnapshot: FilesMetadataSyncSnapshot = {
+    items: useFilesStore.getState().items,
+    libraryState: useFilesStore.getState().libraryState,
+    documents: await serializeIndexedDbStoreSnapshot(STORES.DOCUMENTS),
+    deletedPaths: localDeletedPaths,
+  };
+  const mergedSnapshot = mergeFilesMetadataSnapshots(localSnapshot, {
+    ...data,
+    deletedPaths: remoteDeletedPaths,
+  });
+  const effectiveDeletedPaths = mergeDeletionMarkerMaps(
+    localDeletedPaths,
+    remoteDeletedPaths
+  );
+  const prunedDeletedPaths = Object.keys(effectiveDeletedPaths).filter(
+    (path) => !mergedSnapshot.deletedPaths?.[path]
+  );
+
+  cloudSyncState.mergeDeletedKeys("fileMetadataPaths", remoteDeletedPaths);
+  cloudSyncState.clearDeletedKeys("fileMetadataPaths", prunedDeletedPaths);
+
+  useFilesStore.setState({
+    items: mergedSnapshot.items,
+    libraryState: mergedSnapshot.libraryState,
+  });
+
+  await applyIndexedDbStoreSnapshot(
+    STORES.DOCUMENTS,
+    mergedSnapshot.documents || []
+  );
+}
+
+export const mergeFilesMetadataSyncSnapshots = mergeFilesMetadataSnapshots;
