+++ src/utils/sync/engine/domains/indexedDbStoreSync.ts
@@ -1,0 +1,252 @@
+import { ensureIndexedDBInitialized } from "@/utils/indexedDB";
+
+export interface StoreItem {
+  [key: string]: unknown;
+}
+
+export interface StoreItemWithKey {
+  key: string;
+  value: StoreItem;
+}
+
+export interface SerializedStoreItemRecord {
+  item: StoreItemWithKey;
+  signature: string;
+}
+
+export type FilesStoreSnapshotData = StoreItemWithKey[];
+
+const blobToBase64 = (blob: Blob): Promise<string> =>
+  new Promise((resolve, reject) => {
+    const reader = new FileReader();
+    reader.onloadend = () => resolve(reader.result as string);
+    reader.onerror = () =>
+      reject(reader.error || new Error("Failed to serialize blob"));
+    reader.readAsDataURL(blob);
+  });
+
+const base64ToBlob = (dataUrl: string): Blob => {
+  const [meta, base64] = dataUrl.split(",");
+  const mimeMatch = meta.match(/data:(.*);base64/);
+  const mime = mimeMatch ? mimeMatch[1] : "application/octet-stream";
+  const binary = atob(base64);
+  const array = Uint8Array.from(binary, (char) => char.charCodeAt(0));
+  return new Blob([array], { type: mime });
+};
+
+async function computeSyncSignature(value: unknown): Promise<string> {
+  const payload = new TextEncoder().encode(JSON.stringify(value));
+  const digest = await crypto.subtle.digest("SHA-256", payload);
+  return Array.from(new Uint8Array(digest), (byte) =>
+    byte.toString(16).padStart(2, "0")
+  ).join("");
+}
+
+export async function readStoreItems(
+  db: IDBDatabase,
+  storeName: string
+): Promise<StoreItemWithKey[]> {
+  return new Promise((resolve, reject) => {
+    try {
+      const transaction = db.transaction(storeName, "readonly");
+      const store = transaction.objectStore(storeName);
+      const items: StoreItemWithKey[] = [];
+      const request = store.openCursor();
+
+      request.onsuccess = (event) => {
+        const cursor = (event.target as IDBRequest<IDBCursorWithValue>).result;
+        if (cursor) {
+          items.push({
+            key: cursor.key as string,
+            value: cursor.value as StoreItem,
+          });
+          cursor.continue();
+          return;
+        }
+
+        resolve(items);
+      };
+
+      request.onerror = () => reject(request.error);
+    } catch (error) {
+      reject(error);
+    }
+  });
+}
+
+async function serializeStoreItem(item: StoreItemWithKey): Promise<StoreItemWithKey> {
+  const serializedValue: Record<string, unknown> = {
+    ...item.value,
+  };
+
+  for (const key of Object.keys(item.value)) {
+    if (item.value[key] instanceof Blob) {
+      serializedValue[key] = await blobToBase64(item.value[key] as Blob);
+      serializedValue[`_isBlob_${key}`] = true;
+    }
+  }
+
+  return {
+    key: item.key,
+    value: serializedValue,
+  };
+}
+
+export async function serializeStoreItems(
+  items: StoreItemWithKey[]
+): Promise<StoreItemWithKey[]> {
+  return Promise.all(items.map((item) => serializeStoreItem(item)));
+}
+
+export async function serializeStoreItemRecords(
+  items: StoreItemWithKey[]
+): Promise<SerializedStoreItemRecord[]> {
+  return Promise.all(
+    items.map(async (item) => {
+      const serializedItem = await serializeStoreItem(item);
+      return {
+        item: serializedItem,
+        signature: await computeSyncSignature(serializedItem),
+      };
+    })
+  );
+}
+
+function deserializeStoreItem(item: StoreItemWithKey): Record<string, unknown> {
+  const restoredValue: Record<string, unknown> = {
+    ...item.value,
+  };
+
+  for (const key of Object.keys(item.value)) {
+    const isBlobKey = `_isBlob_${key}`;
+    if (item.value[isBlobKey] === true && typeof item.value[key] === "string") {
+      restoredValue[key] = base64ToBlob(item.value[key] as string);
+      delete restoredValue[isBlobKey];
+    }
+  }
+
+  return restoredValue;
+}
+
+export async function restoreStoreItems(
+  db: IDBDatabase,
+  storeName: string,
+  items: StoreItemWithKey[]
+): Promise<void> {
+  return new Promise((resolve, reject) => {
+    const transaction = db.transaction(storeName, "readwrite");
+    const store = transaction.objectStore(storeName);
+
+    transaction.oncomplete = () => resolve();
+    transaction.onerror = () => reject(transaction.error);
+    transaction.onabort = () =>
+      reject(transaction.error || new Error(`Transaction aborted: ${storeName}`));
+
+    const clearRequest = store.clear();
+
+    clearRequest.onsuccess = () => {
+      try {
+        for (const item of items) {
+          store.put(deserializeStoreItem(item), item.key);
+        }
+      } catch (error) {
+        transaction.abort();
+        reject(error);
+      }
+    };
+
+    clearRequest.onerror = () => reject(clearRequest.error);
+  });
+}
+
+export async function upsertStoreItems(
+  db: IDBDatabase,
+  storeName: string,
+  items: StoreItemWithKey[]
+): Promise<void> {
+  return new Promise((resolve, reject) => {
+    const transaction = db.transaction(storeName, "readwrite");
+    const store = transaction.objectStore(storeName);
+
+    transaction.oncomplete = () => resolve();
+    transaction.onerror = () => reject(transaction.error);
+    transaction.onabort = () =>
+      reject(transaction.error || new Error(`Transaction aborted: ${storeName}`));
+
+    try {
+      for (const item of items) {
+        store.put(deserializeStoreItem(item), item.key);
+      }
+    } catch (error) {
+      transaction.abort();
+      reject(error);
+    }
+  });
+}
+
+export async function deleteStoreItemsByKey(
+  db: IDBDatabase,
+  storeName: string,
+  keys: string[]
+): Promise<void> {
+  if (keys.length === 0) {
+    return;
+  }
+
+  return new Promise((resolve, reject) => {
+    const transaction = db.transaction(storeName, "readwrite");
+    const store = transaction.objectStore(storeName);
+
+    transaction.oncomplete = () => resolve();
+    transaction.onerror = () => reject(transaction.error);
+    transaction.onabort = () =>
+      reject(transaction.error || new Error(`Transaction aborted: ${storeName}`));
+
+    try {
+      for (const key of keys) {
+        store.delete(key);
+      }
+    } catch (error) {
+      transaction.abort();
+      reject(error);
+    }
+  });
+}
+
+export async function serializeIndexedDbStoreSnapshot(
+  storeName: string
+): Promise<FilesStoreSnapshotData> {
+  const db = await ensureIndexedDBInitialized();
+
+  try {
+    const items = await readStoreItems(db, storeName);
+    return await serializeStoreItems(items);
+  } finally {
+    db.close();
+  }
+}
+
+export async function serializeIndexedDbStoreRecords(
+  storeName: string
+): Promise<SerializedStoreItemRecord[]> {
+  const db = await ensureIndexedDBInitialized();
+
+  try {
+    return await serializeStoreItemRecords(await readStoreItems(db, storeName));
+  } finally {
+    db.close();
+  }
+}
+
+export async function applyIndexedDbStoreSnapshot(
+  storeName: string,
+  data: FilesStoreSnapshotData
+): Promise<void> {
+  const db = await ensureIndexedDBInitialized();
+
+  try {
+    await restoreStoreItems(db, storeName, data);
+  } finally {
+    db.close();
+  }
+}
